+++ pubspec.yaml
@@ -3,7 +3,7 @@ description: aXes Controller
 
 publish_to: "none"
 
-version: 1.1.23+172
+version: 1.1.23+177
 
 environment:
   sdk: ">=2.19.3 <3.0.0"
